remove unnecessary variables, remove PlatformLink component, remove useLinkValidation hook and instead move all validation to a newly created validation folder, add additional condition for payload validation, other misc changes such as refactoring and variable renaming
