--- client/src/components/PlatformLink/index.ts
@@ -1,3 +1,0 @@
-import PlatformLink from './PlatformLink';
-
-export default PlatformLink;
