+++ server/src/middleware/payloadValidation.ts
@@ -28,12 +28,20 @@ export const checkLinkPayload = (
     return res.status(400).json({ message: 'Add at least one link' });
   }
 
+  if (links.length > 5) {
+    return res
+      .status(400)
+      .json({ message: 'You cannot add more than 5 links' });
+  }
+
   const allLinksNotEmpty = links.every((link) => link.platform && link.linkUrl);
 
   if (!allLinksNotEmpty)
     return res
       .status(400)
-      .json({ message: 'All links must have a platform and a link url' });
+      .json({
+        message: 'All links must have a unique platform and a link url',
+      });
 
   next();
 };
